# logs: do not copy non matching content when apply `MaskSequences`