@gh123man 2025-02-10
I am slightly concerned about the possible CPU impact of running the rule twice (especially on logs that match frequently/all the time). Curious if you measured this (either via benchmark or a real workload)