@paulcacheux 2025-02-11
@gh123man I added a benchmark to the PR (and its result to the body). To be clear, it will always be possible to create a `mask_sequence` regex that will match often enough to make this PR a regression.

On the other hand, I think the win in the case of a non-matching one is pretty big as well. I see multiple paths forward:
- either we have enough data about how this is used in the wild, and we can justify this change as-is
- we could build some kind of learning mechanism to apply the `regexp.Match` in some cases but not all depending on previous match frequency, but I'm a bit afraid this would be a bit wonky
- we could add a new config field to the processing rules, similar to gcc's builtins "likely/unlikely", allowing customers to opt-in to this early check and keeping the current behavior unchanged by default

Let me know if you have any other idea